# Speed up the Tap#formula_files method