@Bo98 2024-08-25
> I was surprised that `Pathname#extname` was noticeably slower than turning it into a string and checking the end of it.

`Pathname` is pretty bad a lot of the time but in this particular case it may be from https://github.com/Homebrew/brew/blob/17e3bee3dadd9290270248e21f194e4a7000a0b7/Library/Homebrew/extend/pathname.rb#L238-L245

You can test by comparing it to `File.extname`.

`.to_s.ends_with?` is technically incorrect if you had a dotfile named `.rb`, though maybe we don't care too much about that.