@apainintheneck 2024-08-26
Closing in favor of https://github.com/Homebrew/brew/pull/18165